Removed all the added functions from  reviewmodal and put them in main content
